Restart pod containers when ConfigMap has changed
If a user tries to update the values of the relay proxy deployment, the
values will be applied the next time the container restarts. However, it
would be preferable if k8s knew to restart the container automatically
when something has changed.

The [Helm documentation][1] provides a solution for ensuring we
automatically roll the deployment of this container upon ConfigMap
changes. This solution matches the one provided in the initial [GitHub
Issue][2].

[1]: https://helm.sh/docs/howto/charts_tips_and_tricks/#automatically-roll-deployments
[2]: https://github.com/launchdarkly/ld-relay-helm/issues/12