Fix releaser configuration
When building the chart artifacts, I failed to move them out of the
repository directory. This resulted in releaser committing them to the
main branch as part of the merge of the release PR.

This cleans up those incorrectly committed artifacts and fixes the
releaser config so this doesn't happen during the next release.