Fix changing checksum in golden file tests (#17)
We recently added a checksum to the deployment to ensure k8s restarted
the deployment if configuration options have changed. This change
inadvertently caused a failure in our unit tests as the checksum can
change between test runs.

This commit removes the checksum from the generated golden file so we
don't have to worry about the value changing between runs. This is
consistent with how we are handling other values that change like the
Chart version.