Allow mounting secrets as files (#19)
The relay proxy allows specifying a custom TLS certificate. At least one
customer has requested the ability to load this certificate from a k8s
secret.

This commit allows users of this chart to specify secrets in two
flavors-- either to be loaded as environment variables or to be mounted
as a file directly into the container.