Include all previous published versions on GH pages
When releaser is publishing changes to GH pages, it does not attempt to
merge new files in with the old. Instead, it completely replaces that
branch with the files you store in $LD_RELEASE_DOCS_DIR.

Our helm chart needs the newly created archive file, the newly created
index.yaml file, but it also needs all previously published archives,
which this commit addresses.

On the release prior to this one, I restored all previous published
archive files by hand, so going forward this change should be all that
is necessary for correct releases.