feat: Introduce top level pod configuration value (#31)
The chart initially supported a few pod configuration options at the top
level -- podAnnotations and podSecurityContext. Now that we want to
support labels on the pod, we are moving the preferred configuration
under a single pod values option. This new option supports

- pod.annotations :: Takes precedence over podAnnotations
- pod.labels
- pod.securityContext :: Takes precedence over podSecurityContext

The legacy configuration options will continue to be supported until the
next major release.