fix: Allow multiple secrets to be mounted as volumes
When specifying a volume on a container, you cannot share a mountPath.
We were incorrectly trying to do this if more than 1 secret was to be
mounted as a volume.

Our documentation states that all secret volumes will be mounted under
`/mnt/secret`. And indeed, if you attempted to mount a single secret
under the name `key`, you would find the file at `/mnt/secret/key`.

To keep this consistent, we are updating the template to generate a
single projected volume[1] which contains all the secret values, allowing
us to continue mounting multiple secrets relative to `/mnt/secret`.

[1]: https://kubernetes.io/docs/concepts/storage/projected-volumes/