chore: Fix docs publish workflow
Previously, we were cloning the GH pages branch into a separate
repository, updating it with a new index.yaml file, and then trying to
use the launchdarkly/gh-actions/actions/publish-pages action to publish
those modifications.

However, that action assumes the docs directory ONLY contains
publishable files. The .git files in our checkout were conflicting with
the action's own checkout of gh-pages, resulting in a failure.

This update modifies the commands so that we only read from the initial
cloned repository, and operate only within a clean directory.